+++ src/utils/constants.ts
@@ -5,6 +5,8 @@ export default {
     countryCode: '',
     username: '',
     password: '',
+    library: 'stable',
+    verificationCode: '',
     disableDeviceLogging: false,
     devices: [],
     platform: 'Deebot',
@@ -90,6 +92,7 @@ export default {
     pollInterval: 30,
   },
   allowed: {
+    library: ['stable', 'alpha'],
     showAirDryingSwitch: ['presetting', 'yes', 'no'],
   },
   speed2Label: {